packages/yevrah_terminal/requirements.txt
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
cohere>=5.0.0

